service/pyproject.toml
[tool.poetry]
name = "service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.10.1"
torchvision = "^0.11.2"
Pillow = "^9.0.0"
connexion = "^2.10.0"
swagger-ui-bundle = "^0.0.9"
Flask-Cors = "^3.0.10"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-mock = "^3.6.1"
black = {version = "^22.1.0", allow-prereleases = true}
pylint = "^2.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
